requirements.txt
numpy>=1.19.0
opencv-python>=4.5.0
mtcnn>=0.1.0
tensorflow-macos>=2.5.0
tensorflow-metal>=0.5.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
PyQt5>=5.15.0
torch>=1.8.0
torchvision>=0.9.0
tqdm>=4.60.0
Pillow>=8.0.0